+++ Tests/SwifterSwiftUITests/SwifterSwiftUI/SwiftUI/Helpers/TestError.swift
@@ -1,0 +1,12 @@
+//
+//  TestError.swift
+//  SwifterSwiftUITests
+//
+//  Created by Saul Moreno Abril on 02/05/2020.
+//
+
+import Foundation
+
+enum TestError: Error {
+    case test
+}
